# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/api/v1/common/civilities": {
      "get": {
        "operationId": "CommonController_getCivilities",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Récupérer la liste des civilités",
        "tags": [
          "common"
        ]
      }
    },
    "/api/v1/common/species": {
      "get": {
        "operationId": "CommonController_getSpecies",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Récupérer la liste des espèces d'animaux",
        "tags": [
          "common"
        ]
      }
    },
    "/api/v1/clients": {
      "post": {
        "operationId": "ClientController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Client"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Créer un nouveau client",
        "tags": [
          "clients"
        ]
      },
      "get": {
        "operationId": "ClientController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Jean",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PaginatedClientResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Récupérer tous les clients avec pagination et recherche",
        "tags": [
          "clients"
        ]
      }
    },
    "/api/v1/clients/{id}": {
      "get": {
        "operationId": "ClientController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Client"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Récupérer un client par son ID",
        "tags": [
          "clients"
        ]
      },
      "patch": {
        "operationId": "ClientController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Client"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Modifier un client",
        "tags": [
          "clients"
        ]
      },
      "delete": {
        "operationId": "ClientController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Supprimer un client",
        "tags": [
          "clients"
        ]
      }
    },
    "/api/v1/animals": {
      "post": {
        "operationId": "AnimalController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnimalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Animal"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Enregistrer un nouvel animal",
        "tags": [
          "animals"
        ]
      },
      "get": {
        "operationId": "AnimalController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Jean",
              "type": "string"
            }
          },
          {
            "name": "species",
            "required": false,
            "in": "query",
            "description": "Filtrer par espèce d'animal",
            "schema": {
              "type": "string",
              "enum": [
                "chien",
                "chat",
                "hamster",
                "tortue"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PaginatedAnimalResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Récupérer tous les animaux avec pagination et recherche",
        "tags": [
          "animals"
        ]
      }
    },
    "/api/v1/animals/{id}": {
      "get": {
        "operationId": "AnimalController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Animal"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Récupérer un animal par son ID",
        "tags": [
          "animals"
        ]
      },
      "patch": {
        "operationId": "AnimalController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnimalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Animal"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Modifier les informations d'un animal",
        "tags": [
          "animals"
        ]
      },
      "delete": {
        "operationId": "AnimalController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Supprimer un animal",
        "tags": [
          "animals"
        ]
      }
    }
  },
  "info": {
    "title": "VET API",
    "description": "CRM pour vétérinaires",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "vet",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "MetaDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "path"
        ]
      },
      "ResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "civility": {
            "type": "string",
            "enum": [
              "M.",
              "Mme",
              "Mlle"
            ]
          },
          "firstName": {
            "type": "string",
            "example": "Jean"
          },
          "lastName": {
            "type": "string",
            "example": "Dupont"
          },
          "email": {
            "type": "string",
            "example": "jean.dupont@email.com"
          },
          "phone": {
            "type": "string",
            "example": "0612345678"
          },
          "animals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Animal"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "civility",
          "firstName",
          "lastName",
          "email",
          "phone",
          "createdAt",
          "updatedAt"
        ]
      },
      "Animal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "example": "Médor"
          },
          "age": {
            "type": "number",
            "example": 5
          },
          "height": {
            "type": "number",
            "example": 45.5
          },
          "weight": {
            "type": "number",
            "example": 12.2
          },
          "species": {
            "type": "string",
            "enum": [
              "chien",
              "chat",
              "hamster",
              "tortue"
            ]
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "clientId": {
            "type": "number",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "age",
          "height",
          "weight",
          "species",
          "client",
          "clientId",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "number"
          },
          "itemCount": {
            "type": "number"
          },
          "itemsPerPage": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "currentPage": {
            "type": "number"
          }
        },
        "required": [
          "totalItems",
          "itemCount",
          "itemsPerPage",
          "totalPages",
          "currentPage"
        ]
      },
      "PaginatedClientResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Client"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "items",
          "pagination"
        ]
      },
      "CreateClientDto": {
        "type": "object",
        "properties": {
          "civility": {
            "type": "string",
            "enum": [
              "M.",
              "Mme",
              "Mlle"
            ],
            "example": "M."
          },
          "firstName": {
            "type": "string",
            "example": "Jean"
          },
          "lastName": {
            "type": "string",
            "example": "Dupont"
          },
          "email": {
            "type": "string",
            "example": "jean.dupont@email.com"
          },
          "phone": {
            "type": "string",
            "example": "0612345678"
          }
        },
        "required": [
          "civility",
          "firstName",
          "lastName",
          "email",
          "phone"
        ]
      },
      "UpdateClientDto": {
        "type": "object",
        "properties": {
          "civility": {
            "type": "string",
            "enum": [
              "M.",
              "Mme",
              "Mlle"
            ],
            "example": "M."
          },
          "firstName": {
            "type": "string",
            "example": "Jean"
          },
          "lastName": {
            "type": "string",
            "example": "Dupont"
          },
          "email": {
            "type": "string",
            "example": "jean.dupont@email.com"
          },
          "phone": {
            "type": "string",
            "example": "0612345678"
          }
        }
      },
      "PaginatedAnimalResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Animal"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "items",
          "pagination"
        ]
      },
      "CreateAnimalDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Médor"
          },
          "age": {
            "type": "number",
            "example": 3
          },
          "height": {
            "type": "number",
            "example": 45.5
          },
          "weight": {
            "type": "number",
            "example": 12.2
          },
          "species": {
            "type": "string",
            "enum": [
              "chien",
              "chat",
              "hamster",
              "tortue"
            ],
            "example": "chien"
          },
          "clientId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "firstName",
          "age",
          "height",
          "weight",
          "species",
          "clientId"
        ]
      },
      "UpdateAnimalDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Médor"
          },
          "age": {
            "type": "number",
            "example": 3
          },
          "height": {
            "type": "number",
            "example": 45.5
          },
          "weight": {
            "type": "number",
            "example": 12.2
          },
          "species": {
            "type": "string",
            "enum": [
              "chien",
              "chat",
              "hamster",
              "tortue"
            ],
            "example": "chien"
          },
          "clientId": {
            "type": "number",
            "example": 1
          }
        }
      }
    }
  }
}